provide secret after adding slot
